Refactor to consolidate contexts into single export file
