--- client/src/components/product-overview/content-panel/contexts/MissingSkuContext.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const MissingSkuContext = React.createContext(null);
-
-export default MissingSkuContext;
