--- client/src/components/product-overview/content-panel/contexts/SelectedQtyContext.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const SelectedQtyContext = React.createContext(null);
-
-export default SelectedQtyContext;
