--- client/src/components/product-overview/content-panel/contexts/SelectedSkuContext.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const SelectedSkuContext = React.createContext(null);
-
-export default SelectedSkuContext;
