+++ client/src/components/product-overview/content-panel/selectors/AddToCart.jsx
@@ -1,9 +1,7 @@
 import React, { lazy, Suspense, useState, useEffect, useContext } from 'react';
 import styled, { keyframes } from 'styled-components';
 import axios from 'axios';
-import SelectedSkuContext from '../contexts/SelectedSkuContext';
-import SelectedQtyContext from '../contexts/SelectedQtyContext';
-import MissingSkuContext from '../contexts/MissingSkuContext';
+import { SelectedSkuContext, SelectedQtyContext, MissingSkuContext } from '../../contexts';
 const CartModal = lazy(() => import('./CartModal'));
 
 const rotate360 = keyframes`
