+++ client/src/components/product-overview/content-panel/selectors/Size.jsx
@@ -1,8 +1,6 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
-import CurrentStyleContext from '../../contexts/CurrentStyleContext';
-import SelectedSkuContext from '../contexts/SelectedSkuContext';
-import MissingSkuContext from '../contexts/MissingSkuContext';
+import { CurrentStyleContext, SelectedSkuContext, MissingSkuContext } from '../../contexts';
 
 const Button = styled.button`
   margin: 5px;
