--- client/src/components/product-overview/contexts/CurrentStyleContext.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const CurrentStyleContext = React.createContext(null);
-
-export default CurrentStyleContext;
