--- client/src/components/product-overview/contexts/PhotoIndexContext.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const PhotoIndexContext = React.createContext(null);
-
-export default PhotoIndexContext;
