--- client/src/components/product-overview/contexts/PreviewStyleContext.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const PreviewStyleContext = React.createContext(null);
-
-export default PreviewStyleContext;
