+++ client/src/components/product-overview/contexts/index.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export const CurrentStyleContext = React.createContext();
+export const PreviewStyleContext = React.createContext();
+export const PhotoIndexContext = React.createContext();
+export const MissingSkuContext = React.createContext();
+export const SelectedQtyContext = React.createContext();
+export const SelectedSkuContext = React.createContext();
+
+export default {
+  CurrentStyleContext,
+  PreviewStyleContext,
+  PhotoIndexContext,
+  MissingSkuContext,
+  SelectedQtyContext,
+  SelectedSkuContext
+};
